MPD: Increment playlist version on each status response.
I am using ncmpcpp as an MPD client (with groovebasin) and the playlist wasn't
being updated (in ncmpcpp) when I added a new track via a different client
(mpc in this case). It took a bit of figuring but it turns out ncmpcpp
compares the result of a `status` command with the result of the previous
one[1] and only requests updates to the information it thinks has changed[2].
In this case it was using the playlist version number[3] to decide whether to
issue a `plchanges` command.

Although it looks like in the new version it asks for changes in the in the
playlist more often[4] anyway... it aint packaged for debian though. And a new
release hasn't been tagged in a couple of years even though there has been
many changes.

[1]: http://repo.or.cz/w/ncmpcpp.git/blob/6039330c6983bbe7cd7b2f19f872c924cf310567:/src/mpdpp.cpp#l285
[2]: http://repo.or.cz/w/ncmpcpp.git/blob/6039330c6983bbe7cd7b2f19f872c924cf310567:/src/status.cpp#l234
[3]: http://www.musicpd.org/doc/protocol/ch03.html
[4]: http://repo.or.cz/w/ncmpcpp.git/blob/34923a035c51eeec3cd28909aa06c29c637799d5:/src/status.cpp#l178